:necktie: Requset access token with socket
게임 방 입장/게임 시작/게임 결과페이지에서 새로운 access token d 요청
