+++ src/api/core/axios.js
@@ -1,15 +1,11 @@
 import axios from 'axios';
 import { getCookie, setCookie } from '../../utils/cookie';
 import { store } from '../../app/configStore';
+import refresh from './refresh';
 import { toast } from '../../components/toast/ToastProvider';
 
 // 인스턴스 생성
 const instance = axios.create({
-  baseURL: process.env.REACT_APP_API_URL,
-  withCredentials: true,
-});
-
-const refresh = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
   withCredentials: true,
 });
@@ -31,11 +27,6 @@ instance.interceptors.request.use((config) => {
   return config;
 });
 
-refresh.interceptors.request.use((config) => {
-  config.headers.Authorization = null;
-  return config;
-});
-
 // TODO - 로그인 만료 처리 interceptors
 // Unauthorized Error 처리
 instance.interceptors.response.use(
@@ -45,10 +36,14 @@ instance.interceptors.response.use(
   async (error) => {
     const { config } = error;
     if (error.response.status === 401) {
-      const data = await refresh.get('/api/users/issue/token');
-      setCookie(data.headers.authorization);
-      config.headers.Authorization = data.headers.authorization;
+      const response = await refresh.get('/api/users/issue/token');
+      setCookie(response.headers.authorization);
+      config.headers.Authorization = response.headers.authorization;
       return axios(config);
+    }
+    if (error.response.status === 403) {
+      toast.error('로그인이 만료되었습니다.');
+      window.location.href('/login');
     }
   },
 );
